+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+### Dependencies
+
+- Bump Cocoa SDK from v8.9.5 to v8.9.6 ([#2546](https://github.com/getsentry/sentry-dotnet/pull/2546))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#896)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/8.9.5...8.9.6)
+
 ## 3.35.1
 
 ### Fixes